+++ app/features/chatbox/voiceChat.js
@@ -15,7 +15,7 @@ define([
     let isVoiceInit = false;
     let initVoiceChat;
     let initVoiceButton;
-    let btnVoiceChatPic;
+    // let btnVoiceChatPic;
     let stream = null;
     let chunks = [];
     let recorder;
@@ -37,9 +37,9 @@ define([
 
     let btnVoiceChatDescription;
 
-    let cancelRecord;
-    let mouseMoved;
-    let isCanceled = false;
+    // let cancelRecord;
+    // let mouseMoved;
+    // let isCanceled = false;
 
     // const getDuration = (src) => {
     //     return new Promise(function (resolve) {
@@ -90,14 +90,13 @@ define([
             secondCount += 1;
         }, 1000);
 
-        btnVoiceChatDescription.innerHTML = '<button class="cancel-voice-record">❌</button><div>Release button to send</div';
+        btnVoiceChatDescription.innerHTML = `<div><lang data-language="RELEASE_TO_SEND">${GLOBAL.getLangJson().RELEASE_TO_SEND}</lang></div>`;
         btnVoiceChatDescription.style.bottom = '160px';
 
-        cancelRecord = document.querySelector('.cancel-voice-record');
-
+        // cancelRecord = document.querySelector('.cancel-voice-record');
         if (recorder.state === 'inactive') {
             recorder.start();
-            startRecordBtn.style.background = 'red';
+            startRecordBtn.style.background = 'white';
             notiStatus.innerHTML = `
             <div id="bars">
                 <div class="bar"></div>
@@ -116,23 +115,22 @@ define([
 
     const releaseRecord = (event) => {
         pulseRing.classList.remove('active');
-
-        console.log(recorder.state, secondCount);
+        // console.log(recorder.state, secondCount);
 
         if (event === 'mouseup') {
             notiStatus.innerHTML = '';
             notiStatus.textContent = '';
-            btnVoiceChatDescription.innerHTML = 'Hold to speak';
+            btnVoiceChatDescription.innerHTML = `<div><lang data-language="HOLD_TO_SPEAK">${GLOBAL.getLangJson().HOLD_TO_SPEAK}</lang></div>`;
             btnVoiceChatDescription.style.bottom = '10px';
 
-            if (secondCount < 2) {
-                console.log('Too short message');
-            }
+            // if (secondCount < 2) {
+            //     console.log('Too short message');
+            // }
         }
 
         if (event === 'mouseleave') {
             notiStatus.innerHTML = '';
-            btnVoiceChatDescription.innerHTML = 'Hold to speak';
+            btnVoiceChatDescription.innerHTML = `<div><lang data-language="HOLD_TO_SPEAK">${GLOBAL.getLangJson().HOLD_TO_SPEAK}</lang></div>`;
             btnVoiceChatDescription.style.bottom = '10px';
         }
         clearInterval(holdTime);
@@ -145,43 +143,43 @@ define([
         }
     };
 
-    const getMousePosition = (e) => {
-        const clientRect = cancelRecord.getBoundingClientRect();
-        const clientX1 = clientRect.left;
-        const clientX2 = clientRect.right;
-        const clientY1 = clientRect.top;
-        const clientY2 = clientRect.bottom;
-        const checkClientX = e.clientX >= clientX1 && e.clientX <= clientX2;
-        const checkClientY = e.clientY >= clientY1 && e.clientY <= clientY2;
-        if (checkClientX && checkClientY) {
-            console.log('on cancel position');
-            cancelRecord.style.backgroundColor = '#FE8F8F';
-            mouseMoved = true;
-        } else {
-            console.log('on send position');
-            cancelRecord.style.backgroundColor = '#9D9D9D';
-            mouseMoved = false;
-        }
-    };
+    // const getMousePosition = (e) => {
+    //     const clientRect = cancelRecord.getBoundingClientRect();
+    //     const clientX1 = clientRect.left;
+    //     const clientX2 = clientRect.right;
+    //     const clientY1 = clientRect.top;
+    //     const clientY2 = clientRect.bottom;
+    //     const checkClientX = e.clientX >= clientX1 && e.clientX <= clientX2;
+    //     const checkClientY = e.clientY >= clientY1 && e.clientY <= clientY2;
+    //     if (checkClientX && checkClientY) {
+    //         console.log('on cancel position');
+    //         cancelRecord.style.backgroundColor = '#FE8F8F';
+    //         mouseMoved = true;
+    //     } else {
+    //         console.log('on send position');
+    //         cancelRecord.style.backgroundColor = '#9D9D9D';
+    //         mouseMoved = false;
+    //     }
+    // };
 
-    const touchPosition = (e) => {
-        const clientRect = cancelRecord.getBoundingClientRect();
-        const clientX1 = clientRect.left;
-        const clientX2 = clientRect.right;
-        const clientY1 = clientRect.top;
-        const clientY2 = clientRect.bottom;
-        const checkTouchX = e.touches[0].clientX >= clientX1 && e.touches[0].clientX <= clientX2;
-        const checkTouchY = e.touches[0].clientY >= clientY1 && e.touches[0].clientY <= clientY2;
-        if (checkTouchX && checkTouchY) {
-            console.log('on cancel position');
-            cancelRecord.style.backgroundColor = '#FE8F8F';
-            mouseMoved = true;
-        } else {
-            console.log('on send position');
-            cancelRecord.style.backgroundColor = '#9D9D9D';
-            mouseMoved = false;
-        }
-    };
+    // const touchPosition = (e) => {
+    //     const clientRect = cancelRecord.getBoundingClientRect();
+    //     const clientX1 = clientRect.left;
+    //     const clientX2 = clientRect.right;
+    //     const clientY1 = clientRect.top;
+    //     const clientY2 = clientRect.bottom;
+    //     const checkTouchX = e.touches[0].clientX >= clientX1 && e.touches[0].clientX <= clientX2;
+    //     const checkTouchY = e.touches[0].clientY >= clientY1 && e.touches[0].clientY <= clientY2;
+    //     if (checkTouchX && checkTouchY) {
+    //         console.log('on cancel position');
+    //         cancelRecord.style.backgroundColor = '#FE8F8F';
+    //         mouseMoved = true;
+    //     } else {
+    //         console.log('on send position');
+    //         cancelRecord.style.backgroundColor = '#9D9D9D';
+    //         mouseMoved = false;
+    //     }
+    // };
 
     const callAPI = (file, namefile) => {
         const formData = new window.FormData();
@@ -201,16 +199,16 @@ define([
     };
 
     const setMouseUPEvent = () => {
-        console.log('mouseup');
-        if (mouseMoved) {
-            isCanceled = true;
-        } else {
-            isCanceled = false;
-        }
+        // console.log('mouseup');
+        // if (mouseMoved) {
+        //     isCanceled = true;
+        // } else {
+        //     isCanceled = false;
+        // }
         releaseRecord('mouseup');
-        window.removeEventListener('mousemove', getMousePosition);
+        // window.removeEventListener('mousemove', getMousePosition);
         window.removeEventListener('mouseup', setMouseUPEvent);
-        window.removeEventListener('touchmove', touchPosition);
+        // window.removeEventListener('touchmove', touchPosition);
         window.removeEventListener('touchend', setMouseUPEvent);
     };
 
@@ -223,45 +221,49 @@ define([
             } else {
                 mimeTypeBrowser = 'audio/webm';
             }
+
             if (navigator.mediaDevices && navigator.mediaDevices.getUserMedia) {
                 stream = await navigator.mediaDevices.getUserMedia({ audio: true, video: false });
 
+                startRecordBtn.style.background = 'green';
+
                 try {
                     recorder = new window.MediaRecorder(stream, { mimeType: mimeTypeBrowser });
-
                     // Click event for pc
                     startRecordBtn.addEventListener('mousedown', () => {
                         holdRecord();
-                        mouseMoved = false;
-                        window.addEventListener('mousemove', getMousePosition);
+                        // mouseMoved = false;
+                        // window.addEventListener('mousemove', getMousePosition);
                         window.addEventListener('mouseup', setMouseUPEvent);
-                        console.log('MOUSEDOWN');
                     });
 
                     // Touch event for mobile
                     startRecordBtn.addEventListener('touchstart', () => {
                         holdRecord();
-                        mouseMoved = false;
-                        window.addEventListener('touchmove', touchPosition);
-                        window.addEventListener('touchend', setMouseUPEvent);
+                        // mouseMoved = false;
+                        // window.addEventListener('touchmove', touchPosition);
                     });
-                    // startRecordBtn.addEventListener('touchend', () => releaseRecord('mouseup'));
-                    // Touch event
+                    startRecordBtn.addEventListener('touchend', setMouseUPEvent);
 
                     recorder.ondataavailable = e => {
                         chunks.push(e.data);
-                        console.log(e.data);
+                        // console.log(e.data);
                     };
 
                     recorder.onstop = () => {
+                        // console.log(secondCount);
+                        // if (isCanceled || secondCount < 2) {
+                        //     chunks = [];
+                        //     return;
+                        // }
                         console.log(secondCount);
+                        // if (secondCount < 1) {
+                        //     chunks = [];
+                        //     return;
+                        // }
 
-                        if (isCanceled || secondCount < 2) {
-                            chunks = [];
-                            return;
-                        }
-
-                        if (recorder.state === 'inactive' && secondCount >= 2 && !isCanceled) {
+                        if (recorder.state === 'inactive') {
+                            // Remove && !isCanceled
                             const blob = new window.Blob(chunks, { type: mimeTypeBrowser });
                             //  ===  call API ====
                             // postData(blob).then(response => {
@@ -296,6 +298,7 @@ define([
                 }
             } else {
                 console.log('getUserMedia not supported on your browser!');
+                startRecordBtn.style.background = 'yellow';
             }
         }
 
@@ -307,7 +310,14 @@ define([
     const toggleVoiceChat = () => {
         if (!isVoiceInit) {
             initRecordFunc('start');
-            btnVoiceChatPic.src = '/assets/images/keyboard.png';
+            if (GLOBAL.getBodyBgTheme() === 'body_theme_black') {
+                initVoiceChat.innerHTML = '<i class="keyboard-vc-dark"></i>';
+                btnVoiceChatDescription.style.color = '#80C9FF';
+            } else {
+                initVoiceChat.innerHTML = '<i class="keyboard-vc"></i>';
+                btnVoiceChatDescription.style.color = '#111';
+            }
+
             initVoiceButton.style.display = 'block';
             btnVoiceChatDescription.style.display = 'block';
             inputTextChat.classList.add('de-active');
@@ -315,13 +325,13 @@ define([
             isVoiceInit = true;
             initVoiceChat.setAttribute('isUsingVoiceMess', true);
         } else {
-            btnVoiceChatPic.src = '/assets/images/microphone.svg';
+            initRecordFunc('stop');
+            initVoiceChat.innerHTML = '<i class="micro-vc"></i>';
             initVoiceButton.style.display = 'none';
             btnVoiceChatDescription.style.display = 'none';
             inputTextChat.classList.remove('de-active');
             inputTextChat.style.height = '43px';
             attachBtn.style.display = 'block';
-            initRecordFunc('stop');
             isVoiceInit = false;
             initVoiceChat.setAttribute('isUsingVoiceMess', false);
         }
@@ -329,8 +339,7 @@ define([
 
     return {
         onInit: () => {
-            console.log('voice chat');
-            console.log('New test')
+            // console.log('voice chat');
             $progressWrapper = $('.chatbox-progress-upload');
             $pathCircle = $progressWrapper.find('path');
             $percentProgress = $progressWrapper.find('span');
@@ -348,7 +357,7 @@ define([
             attachBtn = document.querySelector('.btn__attach');
 
             btnVoiceChatDescription = document.querySelector('.btn-voice-chat-description');
-            btnVoiceChatPic = document.querySelector('.btn__voice-chat-picture');
+            // btnVoiceChatPic = document.querySelector('.btn__voice-chat-picture');
 
             initVoiceChat.addEventListener('click', toggleVoiceChat);
         }