+++ app/features/login/loginForm.js
@@ -2,12 +2,14 @@ define([
     'axios',
     'shared/functions',
     'app/constant',
-    'features/login/loginERP'
+    'features/login/loginERP',
+    'features/login/loginLagblaster'
 ], (
     axios,
     functions,
     constant,
-    loginERPComp
+    loginERPComp,
+    loginLagblasterComp
 ) => {
     const { setDataToLocalApplication, navigate, getFormData } = functions;
     const {
@@ -25,10 +27,37 @@ define([
     let $loader;
     let $loaderErp;
     let $loginERPForm;
+    let $loginLBForm;
     let loading = false;
     const ob = {};
 
     const leaveERPLoginForm = () => ob.onInit();
+    const leaveLBLoginForm = () => ob.onInit();
+
+    // const checkLBAccountIsExist = (email) => {
+    //     const data = new window.FormData();
+    //     data.append('email', email);
+
+    //     const config = {
+    //         method: 'post',
+    //         url: `${LAGBLASTER_API_BASE}action=p_advanced_user_register_check`,
+    //         data
+    //     };
+
+    //     axios(config)
+    //         .then((response) => {
+    //             console.log(response);
+    //             if (response.message === 'Email is already in use') {
+    //                 setDataToLocalApplication(IS_REGISTERED_LB, 'true');
+    //             } else {
+    //                 setDataToLocalApplication(IS_REGISTERED_LB, 'false');
+    //             }
+    //         })
+    //         .catch((error) => {
+    //             console.log(error);
+    //             alert(error);
+    //         });
+    // };
 
     const onSubmit = (e) => {
         e.preventDefault();
@@ -44,6 +73,7 @@ define([
 
         axios.post(`${BASE_URL}/auth/login`, `password=${password}&email=${email}`).then(res => {
             if (res) {
+                // checkLBAccountIsExist(email);
                 setDataToLocalApplication(SESSION_ID, res.sessionId);
                 setDataToLocalApplication(USER_ID, res.userId);
                 setDataToLocalApplication(ACCESS_TOKEN, res.access_token);
@@ -80,6 +110,15 @@ define([
         });
     };
 
+    const loginLagblaster = () => {
+        console.log('login with LB');
+        $loginForm.hide();
+        $loginERPForm.hide();
+        $loginLBForm.show();
+
+        loginLagblasterComp.onInit('token lagblaster test');
+    };
+
     const loginGoogle = () => {
         const redirectUri = `${window.location.protocol}//${window.location.host}`;
         window.location.assign(`xm/oauth2/authorize/google?redirect_uri=${redirectUri}/oauth2`);
@@ -88,11 +127,18 @@ define([
     ob.onInit = () => {
         loading = false;
         $loginERPForm = $('.erp-login-form');
+        $loginLBForm = $('.lagblaster-login-form');
+
         $loginForm = $('.js_login__form');
         $loginForm.off('submit').on('submit', onSubmit);
+
         $loginForm.find('.erp').off('.loginERP').on('click.loginERP', loginERP);
+        $loginForm.find('.lagblaster').off('.loginLagblaster').on('click.loginLagblaster', loginLagblaster);
         $loginForm.find('.google').off('.loginGoogle').on('click.loginGoogle', loginGoogle);
+
         $loginERPForm.find('.erp-cancel-btn').off('.cancelERP').on('click.cancelERP', leaveERPLoginForm);
+        $loginLBForm.find('.lagblaster-cancel-btn').off('.cancelLBlogin').on('click.cancelLBlogin', leaveLBLoginForm);
+
         $emailField = $loginForm.find('[name="email"]');
         $passwordField = $loginForm.find('[name="password"]');
         $errMess = $loginForm.find('.mess');
@@ -105,6 +151,7 @@ define([
         $loader.hide();
         $loaderErp.hide();
         $loginERPForm.hide();
+        $loginLBForm.hide();
         $loginForm.show();
     };