+++ app/features/modal/lagBlaster/modalLBChangePlan.js
@@ -1,0 +1,22 @@
+define([
+    './modalLBMobileTemplate.js'
+], (
+    modalLBMobileTemplate
+) => {
+    const TITLE = 'Change your plan now';
+    const MODALNAME = 'changePlanModal';
+    const NORMAL_TYPE = 'normal';
+    const CONTENT = `
+        <div class="addMoneyModal-amount">
+            <div class="addMoneyModal-amount__title">CHANGE PLAN NOW</div>
+        </div>  
+        <button>test</button>
+        <div id="test"></div>
+    `;
+
+    return {
+        onInit: () => {
+            modalLBMobileTemplate.onAddModal(TITLE, MODALNAME, CONTENT, NORMAL_TYPE);
+        }
+    };
+});