+++ CHANGELOG.md
@@ -2,6 +2,63 @@
 
 All changes to XM WEB CLIENT project will be documented in this file.
 
+## Version 4
+
+### [4.6] - 2021-10-21
+### Update
+- Update route /meeting.
+- Add comments to refresh token code.
+
+### [4.5] - 2021-10-15
+### Update
+- Improvements to XM meetings.
+- Add 'Accept-Language' header to API requets on interceptor.
+- Improve user upload avatar function.
+### Fix
+- Fix issue on sync method.
+
+### [4.4] - 2021-10-06
+### Add
+- Add new route 'signup'.
+- Handle send and accept invitations on sync method.
+- Add image slider on chats.
+### Update
+- Handle errors properly on sync.js and on request interceptor.
+### Fix
+- Fix issue: handle 'invite_key' when a invited user register a new account.
+
+### [4.3] - 2021-09-29
+### Update
+- Update UI of voice messages recording.
+- Improve response to user on Integrate ERP contacts.
+### Fix
+- It is now possible to forward messages to your own channels.
+
+### [4.2] - 2021-09-24
+### Fix
+- Exclude 'xm/oauth2' requests from sw.js fetch to fix and issue with Google OAuth2.
+
+### [4.1] - 2021-09-24
+### Fix
+- Fixes an issue on scrolling up to see the chat history.
+
+### [4.0] - 2021-09-23
+### Add
+- Now users can send voice messages.
+- It is possible to login with your Google account.
+- Users can create their own channels.
+- Add new route /oauth2 to handle OAuth2 authentication.
+- Show last seen when open a cached chat.
+### Update
+- All XHR request are made with Axios.
+- Add 'Authorization' header with Axios interceptor.
+- Handle refresh of expired tokens.
+- Allow to create groups with one member.
+### Fix
+- Fix error when store chats as offlineData.
+- Fix an error when remove group.
+- Show correct chat options by type of chat.
+
 ## Version 3
 
 ### [3.26] - 2021-09-01